+++ packages/web/src/modules/core/guards/index.ts
@@ -1,6 +1,7 @@
 import { isDevelopEnvironment } from '@/modules/core/environment';
 import { NavigationGuardNext, RouteLocation } from 'vue-router';
 import { usePageStore } from '@/modules/core/store/page.store';
+import { isMaxViewSize } from '@/util';
 
 export const ifDevelopEnvironment = (
   to: RouteLocation,
@@ -20,5 +21,17 @@ export const showMobileNavGuard = (
   next: NavigationGuardNext,
 ): void => {
   usePageStore().showMobileFooter = to.meta.showMobileFooter !== false;
+  next();
+};
+
+export const closeMobileDrawerGuard = (
+  to: RouteLocation,
+  from: RouteLocation,
+  next: NavigationGuardNext,
+): void => {
+  if (isMaxViewSize('sm')) {
+    usePageStore().showSidebar = false;
+  }
+
   next();
 };