+++ packages/web/src/router/index.ts
@@ -16,7 +16,7 @@ import { profileRoute } from '@/modules/profiles/routes/profile-route.util';
 import { usePageStore } from '@/modules/core/store/page.store';
 import { loadProfile } from '@/modules/profiles';
 import { useHelpStore } from '@/modules/help/stores/help.store';
-import { showMobileNavGuard } from '@/modules/core';
+import { closeMobileDrawerGuard, showMobileNavGuard } from '@/modules/core';
 
 const routes: Array<RouteRecordRaw> = [];
 
@@ -89,6 +89,7 @@ router.beforeEach(appConfigGuard);
 router.beforeEach(authGuard);
 router.afterEach(() => usePageStore().setShowAppLoader(false));
 router.beforeEach(showMobileNavGuard);
+router.beforeEach(closeMobileDrawerGuard);
 router.beforeEach((to: RouteLocation, from: RouteLocation, next: NavigationGuardNext) => {
   if (to.query.help === '1') {
     useHelpStore().setShowModal(true);